+++ src/services/expressionChecker.js
@@ -1,18 +1,18 @@
-import { getDiscplLogger } from '../loggingUtil'
-import { AndExpressionChecker } from './andExpressionChecker'
-import { CreateExpressionChecker } from './createExpressionChecker'
-import { EqualExpressionChecker } from './equalExpressionChecker'
-import { IsExpressionChecker } from './isExpressionChecker'
-import { LessThanExpressionChecker } from './lessThanExpressionChecker'
-import { ListExpressionChecker } from './listExpressionChecker'
-import { LiteralExpressionChecker } from './literalExpressionChecker'
-import { MaxExpressionChecker } from './maxExpressionChecker'
-import { MinExpressionChecker } from './minExpressionChecker'
-import { NotExpressionChecker } from './notExpressionChecker'
-import { OrExpressionChecker } from './orExpressionChecker'
-import { ProductExpressionChecker } from './productExpressionChecker'
-import { ProjectionExpressionChecker } from './projectionExpressionChecker'
-import { SumExpressionChecker } from './sumExpressionChecker'
+import { getDiscplLogger } from '../utils/logging_util'
+import { AndExpressionChecker } from '../expressions/andExpressionChecker'
+import { CreateExpressionChecker } from '../expressions/createExpressionChecker'
+import { EqualExpressionChecker } from '../expressions/equalExpressionChecker'
+import { IsExpressionChecker } from '../expressions/isExpressionChecker'
+import { LessThanExpressionChecker } from '../expressions/lessThanExpressionChecker'
+import { ListExpressionChecker } from '../expressions/listExpressionChecker'
+import { LiteralExpressionChecker } from '../expressions/literalExpressionChecker'
+import { MaxExpressionChecker } from '../expressions/maxExpressionChecker'
+import { MinExpressionChecker } from '../expressions/minExpressionChecker'
+import { NotExpressionChecker } from '../expressions/notExpressionChecker'
+import { OrExpressionChecker } from '../expressions/orExpressionChecker'
+import { ProductExpressionChecker } from '../expressions/productExpressionChecker'
+import { ProjectionExpressionChecker } from '../expressions/projectionExpressionChecker'
+import { SumExpressionChecker } from '../expressions/sumExpressionChecker'
 
 export class ExpressionChecker {
   /**
